Fix handling of non-integer half supports
Previously the support was directly for padding but the two end have to be padded by ceil(half_support) each. This causes a problem for the exponential basis function, which has a support of 3. The number of expected knots was set to M + 3 but it should be set to M + 2 * ceil(1.5) which is equal to M + 4.