Fix wrong HermiteSpline eval method.
Eval should only return one value instead of the spline value and the derivative value. The mistake was that the basis function values were not index before multiplying them with the coefficients and tangent values. The basis function eval always returns two values, the actual value of the spline and the derivative value.