Fix handling of hermite splines in fit method.
The problem here was that the tangents values were multiplied with first return value of the basis function instead of the second.
I also had to increase the number of point in the test data from 10 to 20 because the CubicHermite spline with M=5 goes crazy with only 10, i.e. the spline goes to very large values, which causes the test to fail. I assume this happens because the number of points matches the number of coeffs + tangents precisely and we end up over fitting.