Reduce padding by one
Previously the padding extended to ceil(support / 2) however this includes an unnecessary point. For instance if we look at B1 (support = 1), the padding would have been 1 but the value of the basis function of the first knot at  is always zero since  is 0.
Changing the padding to be ceil(support / 2) - 1 removes the superfluous padding knot.